requirements.txt
Pillow==12.0.0
python-dotenv==1.1.1
python-telegram-bot==22.5
Requests==2.32.5
